Move test to tests/utils/
